+++ adlfs/spec.py
@@ -1289,9 +1289,24 @@ class AzureBlobFileSystem(AsyncFileSystem):
                         return True
         except KeyError:
             pass
+        except FileNotFoundError:
+            pass
         try:
-            info = await self._info(path)
-            return info["type"] == "file"
+            container_name, path = self.split_path(path)
+            if not path:
+                # A container can not be a file
+                return False
+            else:
+                async with self.service_client.get_blob_client(
+                    container_name, path
+                ) as bc:
+                    props = await bc.get_blob_properties()
+                try:
+                    if props["metadata"]["is_directory"] == "false":
+                        return True
+                except KeyError:
+                    details = self._details([props])
+                    return details["type"] == "file"
         except:  # noqa: E722
             return False
 
@@ -1300,7 +1315,6 @@ class AzureBlobFileSystem(AsyncFileSystem):
 
     async def _isdir(self, path):
         """Is this entry directory-like?"""
-
         if path in self.dircache:
             for fp in self.dircache[path]:
                 # Files will contain themselves in the cache, but
@@ -1308,8 +1322,27 @@ class AzureBlobFileSystem(AsyncFileSystem):
                 if fp["name"] != path:
                     return True
         try:
-            info = await self._info(path)
-            return info["type"] == "directory"
+            container_name, path = self.split_path(path)
+            if not path:
+                return await self._container_exists(container_name)
+            else:
+                path = path.strip("/")
+                key = path + "/"
+                key_length = len(key)
+                async with self.service_client.get_container_client(
+                    container_name
+                ) as cc:
+                    blob_pages = cc.list_blobs(name_starts_with=key)
+                    async for blob in blob_pages:
+                        if (blob["metadata"].get("is_directory", None) == "true") and (
+                            blob["name"] == path
+                        ):
+                            return True
+                        if (blob.name[:key_length] == key) and (
+                            len(blob["name"]) > len(key)
+                        ):
+                            return True
+                    return False
         except IOError:
             return False