+++ adlfs/tests/test_spec.py
@@ -661,6 +661,7 @@ def test_rm_recursive(storage):
         fs.ls("data/root/c")
 
 
+# @pytest.mark.skip
 def test_mkdir(storage):
     fs = AzureBlobFileSystem(
         account_name=storage.account_name, connection_string=CONN_STR,
@@ -1157,17 +1158,29 @@ def test_put_file(storage):
     fs.rm("putdir", recursive=True)
 
 
-@pytest.mark.skip
 def test_isdir(storage):
     fs = AzureBlobFileSystem(
         account_name=storage.account_name, connection_string=CONN_STR
     )
-    BUCKET = "/name/of/the/bucket"
-    BASE_PATH = BUCKET + "/" + "012345"
-    # EMPTY_DIR = BASE_PATH + "/empty_dir"
+    fs.touch("data/root/a/file.txt")
+    assert fs.isdir("data") is True
+    assert fs.isdir("data/top_file.txt") is False
+    assert fs.isdir("data/root") is True
+    assert fs.isdir("data/root/") is True
+    assert fs.isdir("data/root/rfile.txt") is False
+    assert fs.isdir("data/root/a") is True
+    assert fs.isdir("data/root/a/") is True
 
-    fs.makedirs(BASE_PATH)
-    assert fs.isdir(BASE_PATH) is True
+
+def test_isdir_cache(storage):
+    fs = AzureBlobFileSystem(
+        account_name=storage.account_name, connection_string=CONN_STR
+    )
+    print("checking isdir cache")
+    files = fs.ls("data/root")  # noqa: F841
+    assert fs.isdir("data/root/a") is True
+    assert fs.isdir("data/root/a/") is True
+    assert fs.isdir("data/root/rfile.txt") is False
 
 
 def test_cat(storage):